// src/model/products.model.ts
export interface IFormSearchProducts {
  product_id: string;
  product_name: string;
  product_code: string;
  category_id: string;
  facturers: string;
  status: string;
  price_from: string;
  price_to: string;
  stock_from: string;
  stock_to: string;
  offset: string;
  size: number;
  page: number;
}